--- src/main/java/com/dazzle/shop/faq/FaqService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.faq;
-
-public class FaqService {
-
-}
