--- src/main/java/com/dazzle/shop/faq/impl/FaqServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.faq.impl;
-
-public class FaqServiceImpl {
-
-}
