+++ src/main/java/com/dazzle/shop/model/faq/FaqService.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.faq;
+
+public class FaqService {
+
+}
