+++ src/main/java/com/dazzle/shop/model/faq/impl/FaqDAO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.faq.impl;
+
+public class FaqDAO {
+
+}
