+++ src/main/java/com/dazzle/shop/model/faq/impl/FaqServiceImpl.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.faq.impl;
+
+public class FaqServiceImpl {
+
+}
