+++ src/main/java/com/dazzle/shop/model/order/OrderVO.java
@@ -1,5 +1,20 @@
 package com.dazzle.shop.model.order;
 
-public class OrderVO {
+import java.sql.Timestamp;
 
+import lombok.Data;
+
+@Data
+
+public class OrderVO {
+	private int order_num;
+	private Timestamp order_date;
+	private int delivery_price;
+	private String recipient;
+	private String postal_num;
+	private String address;
+	private String detail_address;
+	private String phone_num;
+	private String request;
+	private String payment;
 }
